fix(codex): compare private host owner IDs safely
Roadmap: Codex app-server generation pool and managed Agent continuity / Phase 2 Broker endpoint pool and private generation host
Contract: C-1 Generation-pinned routing; C-2 Drain and retirement

+++ internal/integrations/agents/codexgenerationhost/host.go
@@ -588,7 +588,7 @@ func ownerPrivateDirectory(path string) (fs.FileInfo, error) {
 		return nil, hostRefuse(HostRefusalPrivateRootInvalid, err)
 	}
 	stat, ok := info.Sys().(*syscall.Stat_t)
-	if !ok || stat.Uid != uint32(os.Geteuid()) {
+	if !ok || int64(stat.Uid) != int64(os.Geteuid()) {
 		return nil, hostRefuse(HostRefusalPrivateRootInvalid, nil)
 	}
 	return info, nil
